requirements.txt
flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
python-dotenv==1.0.0
schedule==1.2.0
lxml==4.9.3
webdriver-manager==4.0.1
